refactor: remove AuthController and ProjectController, update import paths
