+++ src/routes/authRoutes.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import AuthController from '../controllers/AuthController';
+import AuthController from '../controllers/Auth.controller';
 import authMiddleware from '../middleware/authMiddleware';
 
 const router = Router();
